requirements.txt
anyio==4.1.0
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
Deprecated==1.2.14
distlib==0.3.7
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
filelock==3.12.4
fonttools==4.46.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==3.11.0
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.5
vonage==3.11.0
vonage-jwt==1.1.0
wrapt==1.16.0
